Removed all methods from VarianceEstimator to variance_methods directory
