+++ kipet/library/VarianceEstimator.py
@@ -1,38 +1,13 @@
-from __future__ import print_function
-from __future__ import division
-
 import copy
-import os
-import math
-import pdb
-import re
-import six
-import subprocess
-import sys
 import time
-
-import matplotlib.pylab as plt
-from pyomo.core.expr.numvalue import NumericConstant
-from scipy.optimize import least_squares
-from scipy.sparse import coo_matrix
 
 from pyomo.core import *
 from pyomo.dae import *
 from pyomo.environ import *
-from pyomo.opt import (ReaderFactory,
-                       ResultsFormat)
-
-try:
-    from StringIO import StringIO
-except ImportError:
-    from io import StringIO
 
 from kipet.library.Optimizer import *
 from kipet.library.mixins.VisitorMixins import ReplacementVisitor
-from kipet.library.common.objectives import get_objective, conc_objective, spectra_objective
 
-    
-    
 class VarianceEstimator(Optimizer):
     """Optimizer for variance estimation.
 
@@ -150,26 +125,24 @@ class VarianceEstimator(Optimizer):
         num_points = kwds.pop('num_points', None)
         with_plots = kwds.pop('with_plots', False)
         
-        if method not in ['originalchenetal', "direct_sigmas", "alternate"]:
-            method = 'originalchenetal'
-            print("Method not set, so assumed that the Chen method is chosen")
-            
-        if method not in ['originalchenetal', "direct_sigmas", "alternate"]:
-            raise RuntimeError("method must be either \"originalchenetal\", or \"direct_sigmas\", or \"alternate\" ")
+        methods_available = ['originalchenetal', "direct_sigmas", "alternate"]
         
+        if method not in methods_available:
+            method = 'originalchenetal'
+            print(f'Default method of {method} selected')
+            print(f'The available methods are {", ".join(methods_available)}')
+            
         if not self.model.alltime.get_discretization_info():
             raise RuntimeError('apply discretization first before initializing')
 
         self._create_tmp_outputs()
         
-        # deactivates objective functions                
         objectives_map = self.model.component_map(ctype=Objective, active=True)
         active_objectives_names = []
         for obj in objectives_map.values():
-            name = obj.cname()
-            active_objectives_names.append(name)
+            active_objectives_names.append(obj.cname())
             obj.deactivate()
-
+            
         list_components = []
         if species_list is None:
             list_components = [k for k in self._mixture_components]
@@ -189,7 +162,6 @@ class VarianceEstimator(Optimizer):
         
         if hasattr(self.model, 'known_absorbance'):
             warnings.warn("Overriden by species with known absorbance")
-        #############################
             list_components = [k for k in self._mixture_components if k not in self._known_absorbance]
             self._sublist_components = list_components
         
@@ -200,7 +172,8 @@ class VarianceEstimator(Optimizer):
             self.component_set = self._sublist_components
             self.component_var = 'C'
         
-        """inputs section""" # additional section for inputs from trajectory and fixed inputs, CS
+        # Fixed imputs and trajectory section
+        
         self.fixedtraj = fixedtraj
         self.fixedy = fixedy
         self.inputs_sub = inputs_sub
@@ -261,345 +234,24 @@ class VarianceEstimator(Optimizer):
             start = time.time()
              
         if method == 'originalchenetal':  
-            results = self._call_orginal_chen_method(solver, run_opt_kwargs)
+            from kipet.library.variance_methods.chen_method import run_method
+            results = run_method(self, solver, run_opt_kwargs)
             
         elif method == 'alternate':
+            from kipet.library.variance_methods.alternate_method import run_alternate_method
             solver = 'ipopt'
-            results = self._call_alternate_method(solver, run_opt_kwargs)
+            results = run_alternate_method(self, solver, run_opt_kwargs)
         
         elif method == 'direct_sigmas':
-            results = self._call_direct_sigma_method(solver, run_opt_kwargs)
+            from kipet.library.variance_methods.alternate_method import run_direct_sigmas_method
+            results = run_direct_sigma_method(self, solver, run_opt_kwargs)
             
         if report_time:
             end = time.time()
             print("Total execution time in seconds for variance estimation:", end - start)
         
         return results
-    
-    def _call_orginal_chen_method(self, solver, run_opt_kwargs):
-        """This is the original method for estimating variances from Chen et
-        al. 2016
-        
-        """
-        solver_opts = run_opt_kwargs.pop('solver_opts', dict())
-        max_iter = run_opt_kwargs.pop('max_iter', 400)
-        tol = run_opt_kwargs.pop('tolerance', 5.0e-5)
-        tee = run_opt_kwargs.pop('tee', False)
-        norm_order = run_opt_kwargs.pop('norm',np.inf)
-        A = run_opt_kwargs.pop('subset_lambdas', None)
-        init_C = run_opt_kwargs.pop('init_C', None)
-        lsq_ipopt = run_opt_kwargs.pop('lsq_ipopt', False)
-        species_list = run_opt_kwargs.pop('subset_components', None)
-        fixed_device_var = run_opt_kwargs.pop('fixed_device_variance', None)
-        
-        if init_C is None:
-            self._solve_initalization(solver, 
-                                      subset_lambdas=A, 
-                                      solver_opts=solver_opts, 
-                                      tee=tee)
-        else:
-            for t in self._allmeas_times:
-                for k in self._mixture_components:
-                    self.model.C[t, k].value = init_C[k][t]
-                    self.model.Z[t, k].value = init_C[k][t]
 
-            s_array = self._solve_S_from_DC(init_C)
-            S_frame = pd.DataFrame(data=s_array,
-                                   columns=self._mixture_components,
-                                   index=self._meas_lambdas)
-                                                                            
-            if hasattr(self, '_abs_components'):
-                for l in self._meas_lambdas:
-                    for k in self._abs_components:
-                        self.model.S[l, k].value = S_frame[k][l]  # 1e-2
-                        if hasattr(self.model, 'known_absorbance'):
-                            if k in self.model.known_absorbance:
-                                self.model.S[l, k].value = self.model.known_absorbance_data[k][l]
-            else:
-                for l in self._meas_lambdas:
-                    for k in self._mixture_components:
-                        self.model.S[l, k].value = S_frame[k][l]  # 1e-2
-                        if hasattr(self.model, 'known_absorbance'):
-                            if k in self.model.known_absorbance:
-                                self.model.S[l, k].value = self.model.known_absorbance_data[k][l] # ??? [l]
-
-            
-        print("{: >11} {: >20}".format('Iter', '|Zi-Zi+1|'))
-        logiterfile = "iterations.log"
-        if os.path.isfile(logiterfile):
-            os.remove(logiterfile)
-
-        # backup
-        if lsq_ipopt:
-            self._build_s_model()
-            self._build_c_model()
-        else:
-            if species_list is None:
-                self._build_scipy_lsq_arrays()
-            else:
-                lsq_ipopt = True
-                self._build_s_model()
-                self._build_c_model()
-            
-        for it in range(max_iter):
-            
-            rb = ResultsObject()
-            
-            vars_to_load = ['Z', 'C', 'Cs', 'S', 'Y']
-            if not hasattr(self, '_abs_components'):
-                vars_to_load.remove('Cs')
-            rb.load_from_pyomo_model(self.model, to_load=vars_to_load)
-            
-            self._solve_Z(solver)
-
-            if lsq_ipopt:
-                self._solve_S(solver)
-                self._solve_C(solver)
-            else:
-                solved_s = self._solve_s_scipy()
-                solved_c = self._solve_c_scipy()
-                
-            
-            ra=ResultsObject()    
-            vars_to_load = ['Z', 'C', 'Cs', 'S']
-            if not hasattr(self, '_abs_components'):
-                vars_to_load.remove('Cs')
-            ra.load_from_pyomo_model(self.model, to_load=vars_to_load)
-            
-            r_diff = compute_diff_results(rb,ra)
-            Z_norm = r_diff.compute_var_norm('Z', norm_order)
-            
-            if it > 0:
-                print("{: >11} {: >20}".format(it, Z_norm))
-            self._log_iterations(logiterfile, it)
-            if Z_norm < tol and it >= 1:
-                break
-                
-        results = ResultsObject()
-        
-        vars_to_load = ['Z', 'dZdt', 'X', 'dXdt', 'C', 'Cs', 'S', 'Y']
-        if not hasattr(self, '_abs_components'):
-            vars_to_load.remove('Cs')
-        results.load_from_pyomo_model(self.model, to_load=vars_to_load)
-
-        print('Iterative optimization converged. Estimating variances now')
-        solved_variances = self._solve_variances(results, 
-                                                 fixed_dev_var=fixed_device_var)
-        self.compute_D_given_SC(results)
-        
-        param_vals = dict()
-        for name in self.model.parameter_names:
-            param_vals[name] = self.model.P[name].value
-
-        results.P = param_vals
-
-        # removes temporary files. This needs to be changes to work with pyutilib
-        if os.path.exists(self._tmp2):
-            os.remove(self._tmp2)
-        if os.path.exists(self._tmp3):
-            os.remove(self._tmp3)
-        if os.path.exists(self._tmp4):
-            os.remove(self._tmp4)
-            
-        return results
-    
-    def _call_alternate_method(solver, run_opt_kwargs):
-        """Calls the alternative method, whatever that means"""
-        
-        solver_opts = run_opt_kwargs.pop('solver_opts', dict())
-        init_sigmas = run_opt_kwargs.pop('initial_sigmas', float())
-        tee = run_opt_kwargs.pop('tee', False)
-        tol = run_opt_kwargs.pop('tolerance', 5.0e-5)
-        A = run_opt_kwargs.pop('subset_lambdas', None)
-        secant_point2 = run_opt_kwargs.pop('secant_point', None)
-        
-        nu_squared = self.solve_max_device_variance('ipopt',
-                                                    tee=tee,
-                                                    subset_lambdas=A, 
-                                                    solver_opts=solver_opts)
-            
-        if secant_point2 is not None:
-            second_point = secant_point2
-        else:
-            second_point = init_sigmas*10
-        itersigma = dict()
-        
-        itersigma[1] = init_sigmas
-        itersigma[0] = second_point
-        
-        iterdelta = dict()
-        
-        count = 1
-        tol = tol
-        funcval = 1000
-        tee = False
-        
-        iterdelta[0] = self._solve_delta_given_sigma('ipopt',
-                                                     tee=tee, 
-                                                     subset_lambdas=A, 
-                                                     solver_opts=solver_opts, 
-                                                     init_sigmas=itersigma[0])
-
-        while abs(funcval) >= tol:
-            print("Overall sigma value at iteration", count,": ", itersigma[count])
-            
-            new_delta = self._solve_delta_given_sigma('ipopt', 
-                                                      tee=tee, 
-                                                      subset_lambdas=A, 
-                                                      solver_opts=solver_opts, 
-                                                      init_sigmas=itersigma[count])
-            print("New delta_sq val: ", new_delta)
-            
-            iterdelta[count] = new_delta
-            
-            def func1(nu_squared, new_delta, init_sigmas):
-                sigmult = 0
-                nwp = len(self._meas_lambdas)
-                for l in self._meas_lambdas:
-                    for k in self.component_set:
-                       sigmult += value(self.model.S[l, k])
-                funcval = nu_squared - new_delta - init_sigmas*(sigmult/nwp)
-                return funcval, sigmult
-           
-            funcval, sigmult = func1(nu_squared, new_delta, itersigma[count])
-
-            if abs(funcval) <= tol:
-                break
-            else:
-                denom_secant = func1(nu_squared, new_delta, itersigma[count])[0] - func1(nu_squared, iterdelta[count-1], itersigma[count - 1])[0]
-                itersigma[count + 1] = itersigma[count] - func1(nu_squared, new_delta, itersigma[count])[0]*((itersigma[count]-itersigma[count-1])/denom_secant)
-                
-                if itersigma[count + 1] < 0:
-                    itersigma[count + 1] = -1*itersigma[count + 1]
-                count += 1
-                
-        if individual_species:
-            print("Solving for individual species' variance based on the obtained delta")
-            max_likelihood_val, sigma_vals, stop_it, results = \
-                self._solve_sigma_given_delta(solver, 
-                                              subset_lambdas= A, 
-                                              solver_opts=solver_opts, 
-                                              tee=tee, 
-                                              delta=new_delta)
-        else:
-            print("The overall model variance is: ", itersigma[count])
-            #sigma_vals = {k: abs(itersigma[count]) for k in component_set}
-            
-            sigma_vals = {}
-            for k in self.component_set:
-                sigma_vals[k] = abs(itersigma[count])
-            
-        print(f'sigma_vals: {sigma_vals}')
-        results = ResultsObject()
-        
-        vars_to_load = ['Z', 'dZdt', 'X', 'dXdt', 'C', 'Cs', 'S', 'Y']
-        if not hasattr(self, '_abs_components'):
-            vars_to_load.remove('Cs')
-        results.load_from_pyomo_model(self.model, to_load=vars_to_load)
-         
-        results.P = {name: self.model.P[name].value for name in self.model.parameter_names}
-        results.sigma_sq = sigma_vals
-        results.sigma_sq['device'] = new_delta
-    
-        return results
-    
-    def _call_direct_sigmas_method(self, solver, run_opt_kwargs):
-        """Calls the direct sigma method"""
-        
-        solver_opts = run_opt_kwargs.pop('solver_opts', dict())
-        tee = run_opt_kwargs.pop('tee', False)
-        A = run_opt_kwargs.pop('subset_lambdas', None)
-        fixed_device_var = run_opt_kwargs.pop('fixed_device_variance', None)
-        device_range = run_opt_kwargs.pop('device_range', None)
-        num_points = run_opt_kwargs.pop('num_points', None)
-        
-        print("Solving for sigmas assuming known device variances")
-        direct_or_it = "it"
-        
-        if device_range:
-            if not isinstance(device_range, tuple):
-                print("device_range is of type {}".format(type(device_range)))
-                print("It should be a tuple")
-                raise Exception
-            elif device_range[0] > device_range[1]:
-                print("device_range needs to be arranged in order from lowest to highest")
-                raise Exception
-            else:
-                print("Device range means that we will solve iteratively for different delta values in that range")
-        if device_range and not num_points:
-            print("Need to specify the number of points that we wish to evaluate in the device range")
-        if not num_points:
-            pass
-        elif not isinstance(num_points, int):  
-            print("num_points needs to be an integer!")
-            raise Exception
-        if not device_range:
-            device_range = list() 
-        if not device_range and not num_points:
-            direct_or_it = "direct"
-            print("assessing for the value of delta provided")
-            if not fixed_device_var:
-                print("If iterative method not selected then need to provide fixed device variance (delta**2)")
-                raise Exception
-            else:
-                if not isinstance(fixed_device_var, float):
-                    raise Exception("fixed device variance needs to be of type float")
-        
-        if direct_or_it in ["it"]:
-            dist = abs((device_range[1] - device_range[0])/num_points)
-            
-            max_likelihood_vals = []
-            delta_vals = []
-            iteration_counter = []
-            delta = device_range[0]
-            
-            count = 0
-            
-            print('*** Starting Variance Iterations ***')
-           
-            while delta < device_range[1]:
-                
-                print(f"Iteration: {count}\tdelta_sq: {delta}")
-                max_likelihood_val, sigma_vals, stop_it, results = \
-                    self._solve_sigma_given_delta(solver, 
-                                                  subset_lambdas=A, 
-                                                  solver_opts=solver_opts, 
-                                                  tee=tee,
-                                                  delta=delta)
-                
-                if max_likelihood_val >= 5000:
-                    max_likelihood_vals.append(5000)
-                    delta_vals.append(log(delta))
-                    iteration_counter.append(count)
-                    max_likelihood_vals.append(max_likelihood_val)
-                    delta_vals.append(log(delta))
-                else:
-                    iteration_counter.append(count)
-                    
-                delta = delta + dist
-                count += 1 
-
-        else:
-            max_likelihood_val, sigma_vals, stop_it = \
-                self._solve_sigma_given_delta(solver, 
-                                              subset_lambdas=A, 
-                                              solver_opts=solver_opts, 
-                                              tee=tee,
-                                              delta=fixed_device_var)
-   
-        results = ResultsObject()
-        
-        vars_to_load = ['Z', 'dZdt', 'X', 'dXdt', 'C', 'Cs', 'S', 'Y']
-        if not hasattr(self, '_abs_components'):
-            vars_to_load.remove('Cs')
-        results.load_from_pyomo_model(self.model, to_load=vars_to_load)
-        
-        results.P = {name: self.model.P[name].value for name in self.model.parameter_names}
-        results.sigma_sq = sigma_vals
-        results.sigma_sq['device'] = delta
-            
-        return results
-    
     def _warn_if_D_negative(self):
     
         for t in self._meas_times:
@@ -613,1037 +265,6 @@ class VarianceEstimator(Optimizer):
 
         return None
 
-    def _solve_initalization(self, solver, **kwds):
-        """Solves formulation 19 in weifengs paper
-
-           This method is not intended to be used by users directly
-
-        Args:
-            sigma_sq (dict): variances 
-        
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            profile_time (bool,optional): flag to tell pyomo to time the construction and solution of the model. 
-            Default False
-        
-            subset_lambdas (array_like,optional): Set of wavelengths to used in initialization problem 
-            (Weifeng paper). Default all wavelengths.
-
-        Returns:
-
-            None
-
-        """
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', True)
-        set_A = kwds.pop('subset_lambdas', list())
-        profile_time = kwds.pop('profile_time', False)
-        sigmas_sq = kwds.pop('variances', dict())
-
-        if not set_A:
-            set_A = self._meas_lambdas
-        
-        keys = sigmas_sq.keys()
-        # added due to new structure for non_abs species, non-absorbing species not included in S and Cs as subset of C (CS):
-        
-        for k in self.component_set:
-            if k not in keys:
-                sigmas_sq[k] = 0.0
-
-        print("Solving Initialization Problem\n")
-
-        self._warn_if_D_negative()        
-        
-        obj = 0.0
-        for t in self._meas_times:
-            for l in set_A:
-                D_bar = sum(self.model.Z[t, k] * self.model.S[l, k] for k in self.component_set)
-                obj += (self.model.D[t, l] - D_bar) ** 2
-                    
-        self.model.init_objective = Objective(expr=obj)
-        
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        solver_results = opt.solve(self.model,
-                                   tee=tee,
-                                   report_timing=profile_time)
-
-        for t in self._allmeas_times:
-            for k in self._mixture_components:
-                if k in sigmas_sq and sigmas_sq[k] > 0.0:
-                    self.model.C[t, k].value = np.random.normal(self.model.Z[t, k].value, sigmas_sq[k])
-                else:
-                    self.model.C[t, k].value = self.model.Z[t, k].value
-                
-        self.model.del_component('init_objective')
-        
-        return None
-
-    def solve_max_device_variance(self, solver, **kwds):
-        """Solves the maximum likelihood formulation with (C-Z) = 0. solves ntp/2*log(eTe/ntp)
-        and then calculates delta from the solution. See documentation for more details.
-
-        Args:
-            solver_opts (dict, optional): options passed to the nonlinear solver
-        
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            subset_lambdas (array_like,optional): Set of wavelengths to used in initialization problem 
-            (Weifeng paper). Default all wavelengths.
-
-        Returns:
-
-            delta_sq (float): value of the max device variance
-
-        """   
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', True)
-        set_A = kwds.pop('subset_lambdas', list())
-
-        if not set_A:
-            set_A = self._meas_lambdas
-
-        list_components = [k for k in self._mixture_components]
-                    
-        self._sublist_components = list_components
-
-        if hasattr(self.model, 'non_absorbing'):
-            warnings.warn("Overriden by non_absorbing")
-            list_components = [k for k in self._mixture_components if k not in self._non_absorbing]
-            self._sublist_components = list_components
-        
-        if hasattr(self.model, 'known_absorbance'):
-            warnings.warn("Overriden by species with known absorbance")
-            list_components = [k for k in self._mixture_components if k not in self._known_absorbance]
-            self._sublist_components = list_components
-            
-        print("Solving For the worst possible device variance\n")
-        
-        self._warn_if_D_negative()
-        
-        obj = 0.0
-        ntp = len(self._meas_times)
-        nwp = len(self._meas_lambdas)
-        
-        if hasattr(self, '_abs_components'):
-            self.component_set = self._abs_components
-        else:
-            self.component_set = self._sublist_components
-        
-        for t in self._meas_times:
-            for l in set_A:
-                D_bar = sum(self.model.Z[t, k] * self.model.S[l, k] for k in self.component_set)
-                obj += (self.model.D[t, l] - D_bar)**2
-       
-        self.model.init_objective = Objective(expr=obj)
-        
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        solver_results = opt.solve(self.model,
-                                   tee=tee)
-        
-        print("values for the parameters in case with no model variance")
-        for k, v in self.model.P.items():
-            print(k, v.value)
-        
-        etaTeta = 0
-        for t in self._meas_times:
-            for l in set_A:
-                D_bar = sum(value(self.model.Z[t, k]) * value(self.model.S[l, k]) for k in self.component_set)
-                etaTeta += (value(self.model.D[t, l])- D_bar)**2
-
-        deltasq = etaTeta/(ntp*nwp)
-        
-        self.model.del_component('init_objective')
-        print("worst case delta squared: ", deltasq)
-
-        return deltasq
-
-    def _solve_delta_given_sigma(self, solver, **kwds):
-        """Solves the maximum likelihood formulation with fixed sigmas in order to
-        obtain delta. This formulation is highly unstable as the problem is ill-posed
-        with the optimal solution being that C-Z = 0. Should not use.
-
-           This method is not intended to be used by users directly
-
-        Args:
-            init_sigmas (dict): variances 
-        
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            profile_time (bool,optional): flag to tell pyomo to time the construction and solution of the model. 
-            Default False
-        
-            subset_lambdas (array_like,optional): Set of wavelengths to used in initialization problem 
-            (Weifeng paper). Default all wavelengths.
-
-        Returns:
-
-            None
-
-        """   
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', True)
-        set_A = kwds.pop('subset_lambdas', list())
-        profile_time = kwds.pop('profile_time', False)
-        sigmas = kwds.pop('init_sigmas', dict() or float() )
-
-        sigmas_sq = dict()
-        if not set_A:
-            set_A = self._meas_lambdas
-        
-        if isinstance(sigmas, float):
-            for k in self.component_set:
-                sigmas_sq[k] = sigmas
-        
-        elif isinstance(sigmas, dict):
-            keys = sigmas.keys()
-            for k in self.component_set:
-                if k not in keys:
-                    sigmas_sq[k] = sigmas
-           
-        print("Solving delta from given sigmas\n")
-        print(sigmas_sq)
-       
-        self._warn_if_D_negative()  
-       
-        obj = 0.0
-        ntp = len(self._meas_times)
-        nwp = len(self._meas_lambdas) 
-        inlog = 0
-        nc = len(component_set)
-        for t in self._meas_times:
-            for l in set_A:
-                D_bar = sum(self.model.C[t, k] * self.model.S[l, k] for k in self.component_set)
-                #D_bar = sum(self.model.C[t, k] * self.model.S[l, k] for k in self._sublist_components)
-                inlog += (self.model.D[t, l] - D_bar)**2
-        # Orig had sublist in both parts - is this an error?
-
-        # Concentration - correct
-        for t in self._meas_times:
-            for k in self._sublist_components:
-                #obj += conc_objective(self.model, sigma=sigmas_sq)
-                obj += 0.5*((self.model.C[t, k] - self.model.Z[t, k])**2)/(sigmas_sq[k])
-                
-        obj += (nwp)*log((inlog)+1e-12)     
-        self.model.init_objective = Objective(expr=obj)
-        
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        solver_results = opt.solve(self.model,
-                                   tee=tee,
-                                   report_timing=profile_time)
-        residuals = (value(self.model.init_objective))
-        
-        print("residuals: ", residuals)
-
-        for k, v in self.model.P.items():
-            print(k, v.value)
-            
-        etaTeta = 0
-        for t in self._meas_times:
-            for l in set_A:
-                D_bar = sum(value(self.model.C[t, k]) * value(self.model.S[l, k]) for k in self.component_set)
-                etaTeta += (value(self.model.D[t, l]) - D_bar)**2
-        
-        deltasq = etaTeta/(ntp*nwp)  
-        self.model.del_component('init_objective')
-       
-        return deltasq
-    
-    def _solve_sigma_given_delta(self, solver, **kwds):
-        """Solves the maximum likelihood formulation to determine the model variance from a
-        given device variance.
-
-           This method is not intended to be used by users directly
-
-        Args:
-            delta (float): the device variance squared 
-        
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            profile_time (bool,optional): flag to tell pyomo to time the construction and solution of the model. 
-            Default False
-        
-            subset_lambdas (array_like,optional): Set of wavelengths to used in initialization problem 
-            (Weifeng paper). Default all wavelengths.
-            
-            solver_opts (dict, optional): dictionary containing solver options for IPOPT
-
-        Returns:
-
-            residuals (float): the objective function value from the optimization
-            
-            variancesdict (dict): dictionary containing the model variance values
-            
-            stop_it (bool): boolean indicator showing whether no solution was found 
-                                (True) or if there is a solution (False)
-                                
-            solver_results: Variance estimation model results, similar to VarianceEstimator 
-                                        results from previous method
-
-        """   
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', True)
-        set_A = kwds.pop('subset_lambdas', list())
-        profile_time = kwds.pop('profile_time', False)
-        delta = kwds.pop('delta', dict())
-        species_list = kwds.pop('subset_components', None)
-
-        model = self.model.clone()
-
-        if not set_A:
-            set_A = self._meas_lambdas
-            
-        if not self._sublist_components:
-            list_components = []
-            if species_list is None:
-                list_components = [k for k in self._mixture_components]
-                
-            else:
-                for k in species_list:
-                    if k in self._mixture_components:
-                        list_components.append(k)
-                    else:
-                        warnings.warn("Ignored {} since is not a mixture component of the model".format(k))
-    
-            self._sublist_components = list_components
-        
-        self._warn_if_D_negative()  
-        ntp = len(self._meas_times)
-        obj = 0.0
-   
-        for t in self._meas_times:
-            for l in set_A:
-                D_bar = sum(self.model.C[t, k] * self.model.S[l, k] for k in self.component_set)
-                obj += 1/(2*delta)*(self.model.D[t, l] - D_bar)**2
-
-        inlog = {k: 0 for k in self.component_set}
-        self.model.eps = Param(initialize = 1e-8)  
-        
-        variancesdict = {k: 0 for k in self.component_set}
-                
-        for t in self._meas_times:
-            for k in self.component_set:
-                inlog[k] += ((self.model.C[t, k] - self.model.Z[t, k])**2)
-        
-        for k in self.component_set:
-            obj += (ntp/2)*log((inlog[k]/(ntp))+self.model.eps)
-        
-        self.model.init_objective = Objective(expr=obj)
-
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        
-        try:
-            solver_results = opt.solve(self.model,
-                                   tee=tee,
-                                   report_timing=profile_time)
-
-            residuals = (value(self.model.init_objective))
-            for t in self._allmeas_times:
-                for k in self.component_set:
-                    variancesdict[k] += 1 / ntp*((value(self.model.C[t, k]) - value(self.model.Z[t, k]))**2)
-            
-            print("Variances")
-            for k in self.component_set:
-                print(k, variancesdict[k])
-            
-            print("Parameter estimates")
-            for k, v in self.model.P.items():
-                print(k, v.value)
-               
-            stop_it = False
-            
-        except:
-            print("FAILED AT THIS ITERATION")
-            variancesdict = None
-            residuals = 0
-            stop_it = True
-            solver_results = None
-            for k, v in self.model.P.items():
-                print(k, v.value)
-            self.model = m
-            for k, v in self.model.P.items():
-                print(k, v.value)
-        self.model.del_component('init_objective')
-        self.model.del_component('eps')
-
-        return residuals, variancesdict, stop_it, solver_results
-
-    def solve_sigma_given_delta(self, solver, **kwds):
-        """Function that solves for model variances based on a given device variance. Solves
-        The log likelihood function and returns a dictionary containg all sigmas, including
-        the device/delta in order to easily apply to the parameter estimation problem.
-
-           This method is intended to be used by users directly
-
-        Args:
-            solver (str): solver to use to solve the problems (recommended "ipopt")
-            
-            delta (float): the device variance squared 
-        
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            solver_opts (dict,optional): Dictionary containing solver options
-        
-            subset_lambdas (array_like,optional): Set of wavelengths to used in problem 
-            (Weifeng paper). Default all wavelengths.
-
-        Returns:
-
-            all_variances (dict): dictionary containg all sigmas, including the device/delta
-
-        """   
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', True)
-        set_A = kwds.pop('subset_lambdas', list())
-        delta = kwds.pop('delta', dict())
-        
-        residuals, sigma_vals, stop_it, results = self._solve_sigma_given_delta(solver, subset_lambdas= set_A, solver_opts = solver_opts, tee=tee, delta = delta)
-        all_variances = sigma_vals
-        all_variances['device'] = delta
-        return all_variances
-    
-    def _solve_iterative_init(self, solver, **kwds):
-        """This method first fixes params and makes C = Z to solve for S. Requires fixed delta and sigmas.
-        following this, the parameters are freed to users bounds and full problem is solved with fixed variances.
-        This function is only meant to be used as part of a full variance estimation strategy. Currently not
-        implemented in any of the strategies, however it can be useful in future implementations to initialize.
-        
-        Not meant to be directly used by users
-
-        Args:
-            sigma_sq (dict): variances 
-        
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            profile_time (bool,optional): flag to tell pyomo to time the construction and solution of the model. 
-            Default False
-        
-            subset_lambdas (array_like,optional): Set of wavelengths to used in initialization problem 
-            (Weifeng paper). Default all wavelengths.
-            
-            solver_options (dict, optional): Solver options for IPOPT
-
-        Returns:
-
-            None
-
-        """   
-        solver_opts = kwds.pop('solver_options', dict())
-        tee = kwds.pop('tee', True)
-        set_A = kwds.pop('subset_lambdas', list())
-        profile_time = kwds.pop('profile_time', False)
-        sigmas_sq = kwds.pop('variances', dict())
-
-        if not set_A:
-            set_A = self._meas_lambdas
-        
-        keys = sigmas_sq.keys()
-        for k in self.component_set:
-            print(k)
-            if k not in keys:
-                sigmas_sq[k] = 0.0
-        
-        self._warn_if_D_negative()
-        list_components = [k for k in self._mixture_components]
-                 
-        print("Solving Initialization Problem with fixed parameters\n")
-        original_bounds = dict()
-        for v, k in self.model.P.items():
-            low = value(self.model.P[v].lb)
-            high = value(self.model.P[v].ub)
-            original_bounds[v] = (low, high)
-            ub = value(self.model.P[v])
-            lb = ub
-            self.model.P[v].setlb(lb)
-            self.model.P[v].setub(ub)
-       
-        obj = 0.0
-       
-        for t in self._meas_times:
-            for l in set_A:
-                D_bar = sum(self.model.Z[t, k] * self.model.S[l, k] for k in self.component_set)
-                obj += (self.model.D[t, l] - D_bar) ** 2
-        
-        self.model.init_objective = Objective(expr=obj)
-        opt = SolverFactory(solver)
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        solver_results = opt.solve(self.model,
-                                   tee=tee,
-                                   report_timing=profile_time)
-        
-        for k, v in self.model.P.items():
-            print(k, v.value)
-        
-        for t in self._allmeas_times:
-            for k in self._mixture_components:
-                if k in sigmas_sq and sigmas_sq[k] > 0.0:
-                    self.model.C[t, k].value = np.random.normal(self.model.Z[t, k].value, sigmas_sq[k])
-                else:
-                    self.model.C[t, k].value = self.model.Z[t, k].value
-                
-        self.model.del_component('init_objective')
-        
-        for v, k in self.model.P.items():
-            ub = original_bounds[v][1]
-            lb = original_bounds[v][0]
-            self.model.P[v].setlb(lb)
-            self.model.P[v].setub(ub)
-            
-        m = self.model
-
-        m.D_bar = Var(m.meas_times,
-                      m.meas_lambdas)
-        
-        def rule_D_bar(m, t, l):
-            return m.D_bar[t, l] == sum(getattr(m, self.component_var)[t, k] * m.S[l, k] for k in self.component_set)
-        
-        m.D_bar_constraint = Constraint(m.meas_times,
-                                        m.meas_lambdas,
-                                        rule=rule_D_bar)
-
-        # estimation
-        def rule_objective(m):
-            expr = 0
-            for t in m.meas_times:
-                for l in m.meas_lambdas:
-
-                    expr += (m.D[t, l] - m.D_bar[t, l]) ** 2 / (sigmas_sq['device'])
-
-            #expr = spectra_objective(m)
-
-            # This part is not even used????
-            # second_term = 0.0
-            # for t in m.meas_times:
-            #     second_term += sum((m.C[t, k] - m.Z[t, k]) ** 2 / sigmas_sq[k] for k in list_components)
-            return expr
-
-        m.objective = Objective(rule=rule_objective) 
-        
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        
-        print("Solving problem with unfixed parameters")
-        solver_results = opt.solve(self.model,
-                                   tee=tee,
-                                   report_timing=profile_time)
-        
-        for k, v in self.model.P.items():
-            print(k, v.value)      
-
-        m.del_component('objective')
-        self.model = m
-        
-        
-    def _solve_Z(self, solver, **kwds):
-        """Solves formulation 20 in weifengs paper
-
-           This method is not intended to be used by users directly
-
-        Args:
-        
-            solver_opts (dict, optional): options passed to the nonlinear solver
-
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            profile_time (bool,optional): flag to tell pyomo to time the construction and solution of the model. 
-            Default False
-        
-            subset_lambdas (array_like,optional): Set of wavelengths to used in initialization problem 
-            (Weifeng paper). Default all wavelengths.
-
-        Returns:
-
-            None
-
-        """
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', False)
-        profile_time = kwds.pop('profile_time', False)
-        
-        for t in self._allmeas_times:
-            for k in self._sublist_components:
-                if hasattr(self.model, 'non_absorbing'):
-                    if k not in self.model.non_absorbing:
-                        pass
-                    else:
-                        self.model.C[t, k].fixed = True
-
-        obj = 0.0
-        
-        # Conc obj no sigma - will this work with missing data?
-        for k in self._sublist_components:
-            obj += sum((self.model.C[t, k]-self.model.Z[t, k])**2 for t in self._meas_times)
-            
-
-        self.model.z_objective = Objective(expr=obj)
-        if profile_time:
-            print('-----------------Solve_Z--------------------')
-            
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-
-        from pyomo.opt import ProblemFormat
-        solver_results = opt.solve(self.model,
-                                   logfile=self._tmp2,
-                                   tee=tee,
-                                   #show_section_timing=True,
-                                   report_timing=profile_time)
-
-        self.model.del_component('z_objective')
-
-    def _solve_s_scipy(self, **kwds):
-        """Solves formulation 22 in weifengs paper (using scipy least_squares)
-
-           This method is not intended to be used by users directly
-
-        Args:
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            profile_time (bool,optional): flag to tell pyomo to time the construction and solution of the model. 
-            Default False
-
-            ftol (float, optional): Tolerance for termination by the change of the cost function. Default is 1e-8
-
-            xtol (float, optional): Tolerance for termination by the change of the independent variables. Default is 1e-8
-
-            gtol (float, optional): Tolerance for termination by the norm of the gradient. Default is 1e-8.
-
-            loss (str, optional): Determines the loss function. The following keyword values are allowed:
-                'linear' (default) : rho(z) = z. Gives a standard least-squares problem.
-
-                'soft_l1' : rho(z) = 2 * ((1 + z)**0.5 - 1). The smooth approximation of l1 (absolute value) loss. Usually a good choice for robust least squares.
-
-                'huber' : rho(z) = z if z <= 1 else 2*z**0.5 - 1. Works similarly to 'soft_l1'.
-
-                'cauchy' : rho(z) = ln(1 + z). Severely weakens outliers influence, but may cause difficulties in optimization process.
-
-                'arctan' : rho(z) = arctan(z). Limits a maximum loss on a single residual, has properties similar to 'cauchy'.
-            f_scale (float, optional): Value of soft margin between inlier and outlier residuals, default is 1.0
-
-            max_nfev (int, optional): Maximum number of function evaluations before the termination
-
-        Returns:
-            None
-
-        """
-        method = kwds.pop('method','trf')
-        def_tol = 1.4901161193847656e-07
-        ftol = kwds.pop('ftol', def_tol)
-        xtol = kwds.pop('xtol', def_tol)
-        gtol = kwds.pop('gtol', def_tol)
-        x_scale = kwds.pop('x_scale', 1.0)
-        loss = kwds.pop('loss', 'linear')
-        f_scale = kwds.pop('f_scale', 1.0)
-        max_nfev = kwds.pop('max_nfev', None)
-        verbose = kwds.pop('verbose', 2)
-        profile_time = kwds.pop('profile_time', False)
-        tee = kwds.pop('tee', False)
-
-        if profile_time:
-            print('-----------------Solve_S--------------------')
-            t0 = time.time()
-
-        if hasattr(self, '_abs_components'):
-            n = self._nabs_components
-        else:
-            n = self._n_components
-
-        for j, l in enumerate(self._meas_lambdas):
-            for k, c in enumerate(self.component_set):
-                if self.model.S[l, c].value < 0.0 and self._is_D_deriv == False:  #: only less thant zero for non-absorbing
-                    self._s_array[j * n + k] = 1e-2
-                else:
-                    self._s_array[j * n + k] = self.model.S[l, c].value
-
-        for j, t in enumerate(self._allmeas_times):
-            for k, c in enumerate(self.component_set):
-                self._z_array[j * n + k] = self.model.Z[t, c].value
-
-        def F(x, z_array, d_array, nl, nt, nc):
-            diff = np.zeros(nt*nl)
-            for i in range(nt):
-                for j in range(nl):
-                    diff[i*nl+j] = d_array[i, j]-sum(z_array[i*nc+k]*x[j*nc+k] for k in range(nc))
-            return diff
-
-        def JF(x, z_array, d_array, nl, nt, nc):
-            row = []
-            col = []
-            data = []
-            for i in range(nt):
-                for j in range(nl):
-                    for k in range(nc):
-                        row.append(i*nl+j)
-                        col.append(j*nc+k)
-                        data.append(-z_array[i*nc+k])
-            return coo_matrix((data, (row, col)),
-                              shape=(nt*nl, nc*nl))
-
-
-        if self._is_D_deriv == False:
-            lower_bound = 0.0
-        else:
-            lower_bound = -np.inf
-                
-        if tee:
-            res = least_squares(F,
-                                self._s_array,
-                                JF,
-                                (lower_bound ,np.inf),
-                                method,
-                                ftol,
-                                xtol,
-                                gtol,
-                                x_scale,
-                                loss,
-                                f_scale,
-                                max_nfev=max_nfev,
-                                verbose=verbose,
-                                args=(self._z_array,
-                                      self._d_array,
-                                      self._n_meas_lambdas,
-                                      self._n_allmeas_times,
-                                      n)
-                                )
-            
-        else:
-            f = StringIO()
-            with stdout_redirector(f):
-                res = least_squares(F,
-                                    self._s_array,
-                                    JF,
-                                    (lower_bound ,np.inf),
-                                    method,
-                                    ftol,
-                                    xtol,
-                                    gtol,
-                                    x_scale,
-                                    loss,
-                                    f_scale,
-                                    max_nfev=max_nfev,
-                                    verbose=verbose,
-                                    args=(self._z_array,
-                                          self._d_array,
-                                          self._n_meas_lambdas,
-                                          self._n_allmeas_times,
-                                          n)
-                                    )
-                
-            with open(self._tmp3,'w') as tf:
-                tf.write(f.getvalue())
-        
-        if profile_time:
-            t1 = time.time()
-            print("Scipy.optimize.least_squares time={:.3f} seconds".format(t1-t0))
-
-        for j, l in enumerate(self._meas_lambdas):
-            for k, c in enumerate(self.component_set):
-                self.model.S[l, c].value = res.x[j * n + k]
-                if hasattr(self.model, 'known_absorbance'):
-                    if c in self.model.known_absorbance:
-                        self.model.S[l, c].set_value(self.model.known_absorbance_data[c][l])
-
-        return res.success
-
-    def _solve_c_scipy(self, **kwds):
-        """Solves formulation 25 in weifengs paper (using scipy least_squares)
-
-           This method is not intended to be used by users directly
-
-        Args:
-            tee (bool,optional): flag to tell the optimizer whether to stream output
-            to the terminal or not
-        
-            profile_time (bool,optional): flag to tell pyomo to time the construction and solution of the model. 
-            Default False
-
-            ftol (float, optional): Tolerance for termination by the change of the cost function. Default is 1e-8
-
-            xtol (float, optional): Tolerance for termination by the change of the independent variables. Default is 1e-8
-
-            gtol (float, optional): Tolerance for termination by the norm of the gradient. Default is 1e-8.
-
-            loss (str, optional): Determines the loss function. The following keyword values are allowed:
-                'linear' (default) : rho(z) = z. Gives a standard least-squares problem.
-
-                'soft_l1' : rho(z) = 2 * ((1 + z)**0.5 - 1). The smooth approximation of l1 (absolute value) loss. Usually a good choice for robust least squares.
-
-                'huber' : rho(z) = z if z <= 1 else 2*z**0.5 - 1. Works similarly to 'soft_l1'.
-
-                'cauchy' : rho(z) = ln(1 + z). Severely weakens outliers influence, but may cause difficulties in optimization process.
-
-                'arctan' : rho(z) = arctan(z). Limits a maximum loss on a single residual, has properties similar to 'cauchy'.
-            f_scale (float, optional): Value of soft margin between inlier and outlier residuals, default is 1.0
-
-            max_nfev (int, optional): Maximum number of function evaluations before the termination
-
-        Returns:
-            None
-
-        """
-        method = kwds.pop('method','trf')
-        def_tol = 1.4901161193847656e-07
-        ftol = kwds.pop('ftol',def_tol)
-        xtol = kwds.pop('xtol',def_tol)
-        gtol = kwds.pop('gtol',def_tol)
-        x_scale = kwds.pop('x_scale',1.0)
-        loss = kwds.pop('loss','linear')
-        f_scale = kwds.pop('f_scale',1.0)
-        max_nfev = kwds.pop('max_nfev',None)
-        verbose = kwds.pop('verbose',2)
-        profile_time = kwds.pop('profile_time',False)
-        tee =  kwds.pop('tee',False)
-        
-        if profile_time:
-            print('-----------------Solve_C--------------------')
-            t0 = time.time()
-       
-        if hasattr(self, '_abs_components'):
-            n = self._nabs_components
-        else:
-            n = self._n_components
-            
-        for i, t in enumerate(self._allmeas_times):
-            for k, c in enumerate(self.component_set):
-                if getattr(self.model, self.component_var)[t, c].value <= 0.0:
-                    self._c_array[i * n + k] = 1e-15
-                else:
-                    self._c_array[i * n + k] = getattr(self.model, self.component_var)[t, c].value
-
-        for j, l in enumerate(self._meas_lambdas):
-            for k, c in enumerate(self.component_set):
-                self._s_array[j * n + k] = self.model.S[l, c].value
-
-
-        def F(x,s_array,d_array,nl,nt,nc):
-            diff = np.zeros(nt*nl)
-            for i in range(nt):
-                for j in range(nl):
-                    diff[i*nl+j]=d_array[i,j]-sum(s_array[j*nc+k]*x[i*nc+k] for k in range(nc))
-            return diff
-
-        def JF(x,s_array,d_array,nl,nt,nc):
-            row = []
-            col = []
-            data = []
-            for i in range(nt):
-                for j in range(nl):
-                    for k in range(nc):
-                        row.append(i*nl+j)
-                        col.append(i*nc+k)
-                        data.append(-s_array[j*nc+k])
-
-            return coo_matrix((data, (row, col)),
-                              shape=(nt*nl,nc*nt))
-
-        if tee:
-            res = least_squares(F, 
-                                self._c_array, 
-                                JF,
-                                (0.0, np.inf), 
-                                method,
-                                ftol, 
-                                xtol, 
-                                gtol,
-                                x_scale, 
-                                loss, 
-                                f_scale,
-                                max_nfev=max_nfev,
-                                verbose=verbose,
-                                args=(self._s_array,
-                                      self._d_array,
-                                      self._n_meas_lambdas,
-                                      self._n_allmeas_times,
-                                      n)
-                                )
-            
-        else:
-            f = StringIO()
-            with stdout_redirector(f):
-                res = least_squares(F, 
-                                    self._c_array, 
-                                    JF,
-                                    (0.0, np.inf), 
-                                    method,
-                                    ftol, 
-                                    xtol, 
-                                    gtol,
-                                    x_scale, 
-                                    loss, 
-                                    f_scale,
-                                    max_nfev=max_nfev,
-                                    verbose=verbose,
-                                    args=(self._s_array,
-                                          self._d_array,
-                                          self._n_meas_lambdas,
-                                          self._n_allmeas_times,
-                                          n)
-                                    )
-                    
-            with open(self._tmp4, 'w') as tf:
-                tf.write(f.getvalue())
-
-        if profile_time:
-            t1 = time.time()
-            print("Scipy.optimize.least_squares time={:.3f} seconds".format(t1-t0))
-
-        for j, t in enumerate(self._allmeas_times):
-            for k,c in enumerate(self.component_set):
-                getattr(self.model, self.component_var)[t,c].value = res.x[j*n+k]
-       
-        return res.success
-
-    def _solve_variances(self, results, fixed_dev_var = None):
-        """Solves formulation 23 in weifengs paper (using scipy least_squares)
-
-           This method is not intended to be used by users directly
-
-        Args:
-            results (ResultsObject): Data obtained from Weifengs procedure
-            fixed_device_var(float, optional): if we have device variance we input it here
-
-        Returns:
-            bool indicated if variances were estimated succesfully.
-
-        """
-        
-        nl = self._n_meas_lambdas
-        nt = self._n_meas_times
-        b = np.zeros((nl, 1))
-
-        variance_dict = dict()
-        n_val = len(self.component_set)
-            
-        A = np.ones((nl, n_val + 1))
-        reciprocal_nt = 1.0/nt
-        for i, l in enumerate(self._meas_lambdas):
-            for j, t in enumerate(self._meas_times):
-                D_bar = 0.0
-                for w, k in enumerate(self.component_set):
-                    A[i, w] = results.S[k][l]**2
-                    D_bar += results.S[k][l]*results.Z[k][t]
-                b[i] += (self.model.D[t, l] - D_bar)**2
-            b[i] *= reciprocal_nt
-
-        if fixed_dev_var == None:
-            res_lsq = np.linalg.lstsq(A, b, rcond=None)
-            all_nonnegative = True
-            n_vars = n_val + 1
-
-            for i in range(n_vars):
-                if res_lsq[0][i] < 0.0:
-                    if res_lsq[0][i] < -1e-5:
-                        all_nonnegative = False
-                    else:
-                        res_lsq[0][i] = abs(res_lsq[0][i])
-                res_lsq[0][i]
-
-            if not all_nonnegative:
-                x0 = np.zeros(n_val + 1) + 1e-2
-                bb = np.zeros(nl)
-                for i in range(nl):
-                    bb[i] = b[i]
-
-                def F(x, M, rhs):
-                    return rhs - M.dot(x)
-
-                def JF(x, M, rhs):
-                    return -M
-
-                res_lsq = least_squares(F, x0, JF,
-                                        bounds=(0.0, np.inf),
-                                        verbose=2, args=(A, bb))
-                for i, k in enumerate(self.component_set):
-                    variance_dict[k] = res_lsq.x[i]
-                variance_dict['device'] = res_lsq.x[n_val]
-                results.sigma_sq = variance_dict
-                return res_lsq.success
-
-            else:
-                for i, k in enumerate(self.component_set):
-                    variance_dict[k] = res_lsq[0][i][0]
-                variance_dict['device'] = res_lsq[0][n_val][0]
-                results.sigma_sq = variance_dict
-
-        if fixed_dev_var:
-            bp = np.zeros((nl, 1))
-            for i, l in enumerate(self._meas_lambdas):
-                bp[i] = b[i] - fixed_dev_var
-            Ap = np.zeros((nl, n_val))
-            for i, l in enumerate(self._meas_lambdas):
-                for j, t in enumerate(self._meas_times):
-                    for w, k in enumerate(self.component_set):
-                        Ap[i, w] = results.S[k][l] ** 2
-
-            res_lsq = np.linalg.lstsq(Ap, bp, rcond=None)
-            all_nonnegative = True
-            n_vars = n_val
-            for i in range(n_vars):
-                if res_lsq[0][i] < 0.0:
-                    if res_lsq[0][i] < -1e-5:
-                        all_nonnegative = False
-                    else:
-                        res_lsq[0][i] = abs(res_lsq[0][i])
-                res_lsq[0][i]
-
-            for i, k in enumerate(self.component_set):
-                variance_dict[k] = res_lsq[0][i][0]
-            variance_dict['device'] = fixed_dev_var
-            results.sigma_sq = variance_dict
-        
-        return 1
-
-    def _build_scipy_lsq_arrays(self):
-        """Creates arrays for scipy solvers
-
-           This method is not intended to be used by users directly
-
-        Args:
-
-        Returns:
-            None
-
-        """
-        self._d_array = np.zeros((self._n_meas_times, self._n_meas_lambdas))
-        for i, t in enumerate(self._meas_times):
-            for j, l in enumerate(self._meas_lambdas):
-                self._d_array[i,j] = self.model.D[t,l]
-
-        if hasattr(self,'_abs_components'):
-            n_val = self._nabs_components
-        else:
-            n_val = self._n_components
-            
-        self._s_array = np.ones(self._n_meas_lambdas * n_val) 
-        self._z_array = np.ones(self._n_allmeas_times * n_val)
-        self._c_array = np.ones(self._n_allmeas_times * n_val)
- 
-        return None
-    
     def _create_tmp_outputs(self):
         """Creates temporary files for loging solutions of each optimization problem
 
@@ -1667,187 +288,6 @@ class VarianceEstimator(Optimizer):
         
         with open(self._tmp4,'w') as f:
             f.write("temporary file for ipopt output")
-
-    def _log_iterations(self, filename, iteration):
-        """log solution of each subproblem in Weifengs procedure
-
-           This method is not intended to be used by users directly
-
-        Args:
-
-        Returns:
-            None
-
-        """
-        with open(filename, "a") as f:
-            f.write("\n#######################Iteration {}#######################\n".format(iteration))
-            with open(self._tmp2,'r') as tf:
-                f.write("\n#######################Solve Z {}#######################\n".format(iteration))
-                f.write(tf.read())
-            with open(self._tmp3,'r') as tf:
-                f.write("\n#######################Solve S {}#######################\n".format(iteration))
-                f.write(tf.read())
-            with open(self._tmp4,'r') as tf:
-                f.write("\n#######################Solve C {}#######################\n".format(iteration))
-                f.write(tf.read())
-
-    def _build_s_model(self):
-        """Builds s_model to solve formulation 22 with ipopt
-
-           This method is not intended to be used by users directly
-
-        Args:
-
-        Returns:
-            None
-
-        """
-        self.S_model = ConcreteModel()
-        if self._is_D_deriv:
-            lower_bound = 0.0
-        else:
-            lower_bound = None
-            
-        self.S_model.S = Var(self._meas_lambdas,
-                             self.component_set,
-                             bounds=(lower_bound, None),
-                             initialize=1.0)
-    
-        for l in self._meas_lambdas:
-            for k in self.component_set:
-                self.S_model.S[l, k].value = self.model.S[l, k].value
-                if hasattr(self.model, 'known_absorbance'):
-                    if k in self.model.known_absorbance:
-                        if self.model.S[l, k].value != self.model.known_absorbance_data[k][l]:
-                            self.model.S[l, k].set_value(self.model.known_absorbance_data[k][l])
-                            self.S_model.S[l, k].fix()
-
-    def _solve_S(self, solver, **kwds):
-        """Solves formulation 23 from Weifengs procedure with ipopt
-
-           This method is not intended to be used by users directly
-
-        Args:
-
-        Returns:
-            None
-
-        """
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', False)
-        update_nl = kwds.pop('update_nl', False)
-        profile_time = kwds.pop('profile_time', False)
-            
-        for l in self._meas_lambdas:
-            for c in self.component_set:
-                self.S_model.S[l, c].value = self.model.S[l, c].value
-                if hasattr(self.model, 'known_absorbance'):
-                    if c in self.model.known_absorbance:
-                        if self.model.S[l, c].value != self.model.known_absorbance_data[c][l]:
-                            self.model.S[l, c].set_value(self.model.known_absorbance_data[c][l])
-                            self.S_model.S[l, c].fix()
-        
-        #D_bar obj
-        obj = 0.0
-        for t in self._meas_times:
-            for l in self._meas_lambdas:
-                D_bar = sum(self.S_model.S[l, k] * self.model.Z[t, k].value for k in self.component_set)
-                obj += (D_bar - self.model.D[t, l]) ** 2
-                    
-        self.S_model.objective = Objective(expr=obj)
-
-        if profile_time:
-            print('-----------------Solve_S--------------------')
-
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        solver_results = opt.solve(self.S_model,
-                                   logfile=self._tmp3,
-                                   tee=tee,
-                                   #keepfiles=True,
-                                   #show_section_timing=True,
-                                   report_timing=profile_time)
-
-        self.S_model.del_component('objective')
-        
-        for l in self._meas_lambdas:
-            for c in self.component_set:
-                self.model.S[l, c].value = self.S_model.S[l, c].value
-                if hasattr(self.model, 'known_absorbance'):
-                    if c in self.model.known_absorbance:
-                        if self.model.S[l, c].value != self.model.known_absorbance_data[c][l]:
-                            self.model.S[l, c].set_value(self.model.known_absorbance_data[c][l])
-                            self.S_model.S[l, c].fix()
-                   
-    def _build_c_model(self):
-        """Builds s_model to solve formulation 25 with ipopt
-
-           This method is not intended to be used by users directly
-
-        Args:
-
-        Returns:
-            None
-
-        """
-        self.C_model = ConcreteModel()
-        self.C_model.C = Var(self._allmeas_times,
-                             self._sublist_components,
-                             bounds=(0.0, None),
-                             initialize=1.0)
-
-        for l in self._allmeas_times:
-            for k in self._sublist_components:
-                self.C_model.C[l, k].value = self.model.C[l, k].value
-                if hasattr(self.model, 'non_absorbing'):
-                    self.C_model.C[l, k].fix()
-        
-    def _solve_C(self,solver,**kwds):
-        """Solves formulation 23 from Weifengs procedure with ipopt
-
-           This method is not intended to be used by users directly
-
-        Args:
-
-        Returns:
-            None
-
-        """
-        solver_opts = kwds.pop('solver_opts', dict())
-        tee = kwds.pop('tee', False)
-        update_nl = kwds.pop('update_nl', False)
-        profile_time = kwds.pop('profile_time', False)
-
-        # D_bar obj
-        obj = 0.0
-        for t in self._meas_times:
-            for l in self._meas_lambdas:
-                D_bar = sum(self.model.S[l, k].value*self.C_model.C[t, k] for k in self.component_set)
-                obj += (self.model.D[t, l]-D_bar)**2
-        
-        self.C_model.objective = Objective(expr=obj)
-                
-        if profile_time:
-            print('-----------------Solve_C--------------------')
-
-        opt = SolverFactory(solver)
-
-        for key, val in solver_opts.items():
-            opt.options[key]=val
-        solver_results = opt.solve(self.C_model,
-                                   logfile=self._tmp4,
-                                   tee=tee,
-                                   #keepfiles=True,
-                                   report_timing=profile_time)
-
-        self.C_model.del_component('objective')
-        
-        for t in self._allmeas_times:
-            for c in self.component_set:
-                self.model.C[t, c].value = self.C_model.C[t, c].value
-         
     
     def load_discrete_jump(self):
         self.jump = True
@@ -1857,7 +297,7 @@ class VarianceEstimator(Optimizer):
             zeit = i
             break
         if zeit is None:
-            raise Exception('no continuous_set')
+            raise Exception('No continuous_set')
 
         self.time_set = zeit.name
 
@@ -1950,7 +390,7 @@ class VarianceEstimator(Optimizer):
                                         con[idx].set_value(e_new)
                                         conlist.add(con[idx].expr)
                     kn = kn + 1
-#############################################################
+
 
 def add_warm_start_suffixes(model):
     # Ipopt bound multipliers (obtained from solution)
@@ -1961,15 +401,6 @@ def add_warm_start_suffixes(model):
     model.ipopt_zU_in = Suffix(direction=Suffix.EXPORT)
     # Obtain dual solutions from first solve and send to warm start
     model.dual = Suffix(direction=Suffix.IMPORT_EXPORT)
-
-
-def compute_diff_results(results1,results2):
-    diff_results = ResultsObject()
-    diff_results.Z = results1.Z - results2.Z
-    diff_results.S = results1.S - results2.S
-    diff_results.C = results1.C - results2.C
-    return diff_results
-
 
 def t_ij(time_set, i, j):
     # type: (ContinuousSet, int, int) -> float
